pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"
faker = "^30.8.2"
asyncpg = "^0.30.0"
pydantic-settings = "^2.6.1"
fastapi = "^0.115.4"
greenlet = "^3.1.1"
python-jose = "^3.3.0"
redis = {extras = ["hiredis"], version = "^5.2.0"}
alembic = "^1.14.0"
uvicorn = "^0.32.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
time-machine = "^2.16.0"
fakeredis = "^2.26.1"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.2"

[tool.ruff]
ignore = ["E501"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
